rename SideBarComponent to sidebar-layout
